feat: implement markdown rendering and structured response components for chatbot
