+++ next/components/organisms/chatbot/shimmer.js
@@ -1,0 +1,6 @@
+import { keyframes } from "@emotion/react";
+
+export const pensandoTextShimmer = keyframes`
+  0%, 100% { color: #6B7280; }
+  50% { color: #252A32; }
+`;
